Move ProcessState over to process manager
